+++ resource_estimator/src/estimates/error_budget.rs
