+++ resource_estimator/src/estimates/layout.rs
