+++ resource_estimator/src/estimates/optimization.rs
@@ -1,8 +1,5 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-mod code_distance_iterators;
-mod distillation_units_map;
-mod tfactory_exhaustive;
-
-pub(crate) use tfactory_exhaustive::TFactoryBuilder;
+mod population;
+pub use population::{Point, Point2D, Point4D, Population};
