--- resource_estimator/src/estimates2/optimization.rs
@@ -1,5 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-mod population;
-pub use population::{Point, Point2D, Point4D, Population};
