+++ resource_estimator/src/system/data/logical_counts.rs
