+++ resource_estimator/src/system/data/physical_counts.rs
