+++ resource_estimator/src/system/serialization.rs
