+++ resource_estimator/src/system/stages.rs
