+++ resource_estimator/src/system/stages/layout.rs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-use crate::estimates2::Overhead;
+use crate::estimates::Overhead;
 
 use super::super::{
     constants::{
